Chat: moved the Kabelwerk-related interfaces to a separate file.
