+++ src/chat/KabelwerkContext.tsx
@@ -4,64 +4,9 @@ import { ReactElement } from 'react';
 import { useSelector } from 'react-redux';
 import { isAuthenticated as isAuthenticatedSelector } from '../auth/authReducer';
 import { getChatAuthorizationToken } from '../utils/api';
-import { INBOXES } from './inboxes';
-import { InboxType } from './pages/Inbox';
+import { INBOXES, InboxType } from './inboxes';
 import useNotification from './hooks/useNotification';
-
-export interface Inbox {
-  loadMore: () => Promise<{ items: InboxItem[] }>;
-}
-
-export interface InboxItem {
-  room: {
-    archived: boolean;
-    assignedTo: number | null;
-    attributes: any;
-    id: number;
-    user: User;
-  };
-  message: Message | null;
-  isNew: boolean;
-}
-
-export interface Room {
-  loadEarlier: () => Promise<{ messages: Message[] }>;
-  isArchived: () => boolean;
-  postMessage: (message: { text: string }) => Promise<Message>;
-  off: () => void;
-  archive: () => Promise<void>;
-  unarchive: () => Promise<void>;
-  getUser: () => User;
-  updateHubUser: (userId: number | null) => Promise<void>;
-  getHubUser: () => User;
-}
-
-export enum MessageType {
-  Text = 'text',
-  RoomMove = 'room_move',
-}
-
-export interface Message {
-  id: number;
-  insertedAt: Date;
-  roomId: number;
-  text: string;
-  type: MessageType;
-  updatedAt: Date;
-  user: User | null;
-}
-
-interface HubInfo {
-  id: number;
-  name: string;
-  users: User[];
-}
-
-interface User {
-  id: number;
-  key: string;
-  name: string;
-}
+import { HubInfo, User, Inbox, InboxItem, Room, Message } from './interfaces';
 
 export const KabelwerkContext = React.createContext<{
   connected: boolean;
